Refactor resource converter state.
